mv abc
